requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=1.6
Babel>=1.3
stevedore>=1.5.0 # Apache-2.0
six>=1.9.0
iso8601>=0.1.9
zake>=0.1.6 # Apache-2.0
msgpack-python>=0.4.0
fasteners>=0.7 # Apache-2.0
retrying!=1.3.0,>=1.2.3 # Apache-2.0
futures>=3.0;python_version=='2.7' or python_version=='2.6'
futurist>=0.1.2 # Apache-2.0
oslo.utils!=2.6.0,>=2.4.0 # Apache-2.0
oslo.serialization>=1.10.0 # Apache-2.0
python-etcd>=0.4.2<0.5.0
